+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-06-29
+
+### Changed
+
+- New field `cet_applies_until_trade_remedy_transition_reviews_concluded` added to tariff dataset
+
 ## 2020-06-23
 
 ### Changed